Refactored views.py and created dedicated files according to the function
