+++ backend/.gitignore
@@ -49,3 +49,5 @@ yarn-error.log*
 # Misc
 .cache/
 .temp/
+
+Think41data.pdf
